fix(ol-input): read the microphone status at its real width
authorizationStatusForMediaType: returns an NSInteger, so declaring the
return as i32 made every debug build panic on objc2's message check the
first time permissionStatus() was called. Widening it to isize also opens
the keyboard and event-source primitives to the modules that reuse them
rather than writing a second input path.

+++ native/ol-input/src/platform/linux.rs
@@ -189,7 +189,7 @@ pub fn request_accessibility() -> bool {
 
 /// PipeWire and PulseAudio grant microphone access at the session level, so
 /// there is nothing to probe: the capture attempt is the probe.
-pub fn microphone_status() -> i32 {
+pub fn microphone_status() -> isize {
     3
 }
